Refactor: ThirdStepAdd 제거 및 타입 통합
